make <localleader>nt map available in non-GUI agenda
This may only be part of the fix. For some reason some mappings are done
at the end of ftplugin/org.vim. These are unavailable in the agenda
view.

I did not thoroughly check, but this seems to be the most important
missing binding for terminal support in the agenda view.